pushing progress in case my laptop spontaneously combusts (WIP)
First problem of extra concat is solved, but now the problem is that the head offset is off by one byte (0x0024 with bool, 0x0025 with uint8). I think the problem is that the initial head offset is calculated before processing the boolean values, resulting in them not being accounted for in the offset.